Add new python code
This Python script implements a simple "Odd or Even" game where a user plays against the computer. The user first chooses either "odd" or "even" and then inputs a number between 1 and 6. The computer also randomly selects a number in the same range, and the sum of the two numbers determines who gets to "bat." If the sum matches the user's choice (odd or even), the user bats; otherwise, the computer bats. Each player aims to avoid getting "out," which occurs if both the user and computer choose the same number. The game continues until one of them is "out," determining the winner.